Add initial piece table implementation.
It's apt to be very slow since we deal with only characters, and read
files one character at a time, etc. This is just to match what's in the
paper -- it shouldn't be hard to generalize to read and write strings.

I haven't tested it except for reading a whole file, which only makes
one piece.