Mark motions as linewise/inclusive/exclusive.
This affects the behavior of operators like this:

* If an operator is applied to a linewise motion, the affected region is
extended to include the entire start and end lines, not just the
characters between the start and end character.

* If an operator is applied to an exclusive motions, the last character
of the region is not affected. e.g. 'w' is exclusive. 'w' jumps to the first
character of the next word, but this first character is not deleted by
'dw'. On the other hand, 'e' is exclusive. 'e' jumps to the last
character of the next word, and 'de' does delete this character.