Fix bug in buf_insert.
We were computing the new capacity directly in buf->cap, but this isn't
OK to do since buf_grow checks whether buf->cap is big enough and skips
the realloc if it is, which means that we were just increasing the cap
without actually making room.

The bug manifested as a segfault when trying to yank after having done a
search, presumably because the '/' and '"' registers are next to each
other in memory.