Refactoring: have all edits go through the buffer.
Instead of having code call gb_put* or gb_del directly, have it call
the new buffer_do_insert and buffer_do_delete functions so we have one
place to process all edits to a buffer.

Reshuffle the undo code around to take advantage of this. In the future,
we can also update a buffer's marks inside these functions.