Fix memory leak when jumping to a tag.
If we navigate down the tag stack with C-t and then jump to a tag, the
behavior is to clear all the tags above the current one and then append
the new tag. Previously we didn't free the tags in that case, just
manipulated the prev & next pointers to skip over them.