Generalize splits.
* Support vertical and horizontal splits (adding :split).
* Add splitbelow option, similar to splitright.
* Navigate between splits with <C-w>+<C-{h,j,k,l}> or <C-w>+{h,j,k,l}.
* Resize a split with <C-w>{<,>,-,+}, prefixed with optional count.

The window layout is represented as a binary tree where the internal
nodes are splits and the leaves are windows. We only store the width and
height of the root window (i.e. the screen), and split offsets and
directions for each split. From this we can compute the width, height,
and x & y positions for any particular window by walking up the tree.