Fix parallel builds.
Specifying mutiple targets in a rule doesn't indicate that the rule
produces multiple outputs -- it actually defines a separate but
identical rule for each of the targets.

So parallel builds were broken because termbox was being built twice in
parallel in order to separately satisfy the 'termbox.h' and
'libtermbox.a' targets, and this caused waf to fail.

Fix by defining a rule for the termbox build directory, and having
termbox.h and libtermbox.a depend on that.

A nice article on makefile rules with multiple outputs is here:
https://www.cmcrossroads.com/article/rules-multiple-outputs-gnu-make