Refactor window_split to take a direction.
This gives the caller more control over the splitting, instead of the
behavior implicitly depending on the values of the 'splitbelow' and
'splitright' options. This lets us simplify the -o and -O flag handling,
and also makes the tests clearer.

This is also a step towards moving the options onto the editor struct --
right now they are global variables for no good reason.