Support window-local and buffer-local options.
Some options, including some that we've already implemented, are meant
to be local to a window or to a buffer rather than global.

Split windows inherit the window-local options of their parents.

Buffer-local options actually have both a global and a local value --
the global value is used to initialize new buffers. :set sets both the
global and local values -- :setlocal and :setglobal (new in this commit)
set only one of them.

Make number, relativenumber, numberwidth & cursorline window-local.

Along the way, this commit refactors the way options are implemented.
Previously options were always accessed by looking them up by name in a
table and casting the value to the appropriate type. Now code can simply
access the fields on the appropriate struct. There's no need for the
caller to re-specify the type, and a typo in the option name is a
compilation error rather than runtime assert failure. There are still
tables to power :set / :setl / :setg.

(This commit declares 'modifiable' as an example of a buffer-local
option to avoid 'unused macro' warnings. The implementation will come in
a later commit.)