Refactor editor commands.
Replace the static table of commands with a linked list, and register
commands via the new EDITOR_COMMAND macro, which will automatically add
the command to the list using a constructor function.

Use this opportunity to move the implementations of some commands out of
editor.c where it makes sense (for example, move :set to options.c).

Also generalize the handling of "!" in commands. Rather than being part
of the command name, it's recognized by editor_execute_command and
translated into a 'force' boolean argument.