Support commands with multiline output.
Add a new "message" buf on the editor. If this is non-empty, it's drawn
between the root window and the status bar.

The effect is that the message appears to "push up" the window so that
it extends beyond the top of the screen. This is achieved by drawing the
window as normal, then shifting the bottom cells up (using memmove)
before drawing the message.