Clean up termbox mock using symbol interposition.
Instead of not linking in termbox at all in the tests, link it in but
but link it dynamically. This lets us override only a few of the
functions and use the rest. In particular, thanks to symbol
interposition, tb_string_with_limit, tb_string, tb_stringf and tb_empty
all end up calling our version of tb_char.